fix: avoid circular imports with models to house BaseModel classes
